# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/brokers": {
      "get": {
        "operationId": "BrokersController_findAll",
        "summary": "Mirrors BrokerRepository.all().",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BrokerDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "brokers"
        ]
      },
      "post": {
        "operationId": "BrokersController_create",
        "summary": "B01 — create the caller’s broker profile. Starts unverified.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrokerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrokerDto"
                }
              }
            }
          }
        },
        "tags": [
          "brokers"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/properties": {
      "get": {
        "operationId": "PropertiesController_findAll",
        "summary": "Mirrors PropertyRepository.all()/.discoverable() — pass discoverableOnly=true for the latter.",
        "parameters": [
          {
            "name": "discoverableOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertyDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ]
      },
      "post": {
        "operationId": "PropertiesController_create",
        "summary": "Publish a listing under the caller’s own broker profile (B03 save).",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDto"
                }
              }
            }
          }
        },
        "tags": [
          "properties"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/search/brokers": {
      "get": {
        "operationId": "SearchController_findBrokers",
        "summary": "Mirrors DiscoveryService.findBrokers, ranked server-side via PostGIS.",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "Client latitude — required to rank/sort by distance.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "description": "Client longitude — required to rank/sort by distance.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "RENT",
                "SALE"
              ],
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "APARTMENT",
                "HOUSE",
                "LAND",
                "STUDIO",
                "ROOM"
              ],
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radiusMeters",
            "required": false,
            "in": "query",
            "description": "Meters.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Free text against broker name/coverage or property title/neighbourhood.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrokerSearchResultsDto"
                }
              }
            }
          }
        },
        "tags": [
          "search"
        ]
      }
    },
    "/search/properties": {
      "get": {
        "operationId": "SearchController_findProperties",
        "summary": "Mirrors DiscoveryService.findProperties, sorted by distance.",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "Client latitude — required to rank/sort by distance.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "description": "Client longitude — required to rank/sort by distance.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "RENT",
                "SALE"
              ],
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "APARTMENT",
                "HOUSE",
                "LAND",
                "STUDIO",
                "ROOM"
              ],
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radiusMeters",
            "required": false,
            "in": "query",
            "description": "Meters.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Free text against broker name/coverage or property title/neighbourhood.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertySearchResultsDto"
                }
              }
            }
          }
        },
        "tags": [
          "search"
        ]
      }
    },
    "/search/suggestions": {
      "get": {
        "operationId": "SearchController_suggestions",
        "summary": "Returns ranked server-side search suggestions.",
        "parameters": [
          {
            "name": "lat",
            "required": true,
            "in": "query",
            "description": "Client latitude — required to rank/sort by distance.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "required": true,
            "in": "query",
            "description": "Client longitude — required to rank/sort by distance.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "transaction",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "RENT",
                "SALE"
              ],
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "APARTMENT",
                "HOUSE",
                "LAND",
                "STUDIO",
                "ROOM"
              ],
              "type": "string"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "radiusMeters",
            "required": false,
            "in": "query",
            "description": "Meters.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Free text against broker name/coverage or property title/neighbourhood.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchSuggestionsDto"
                }
              }
            }
          }
        },
        "tags": [
          "search"
        ]
      }
    },
...
  },
  "info": {
    "title": "Woutalma Keur API",
    "description": "Phase 1: broker/property read, ranked search, contact logging, review eligibility.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "GeoPointDto": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          }
        },
        "required": [
          "latitude",
          "longitude"
        ]
      },
      "BrokerDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "INDIVIDUAL",
              "AGENCY"
            ]
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string",
            "nullable": true
          },
          "position": {
            "$ref": "#/components/schemas/GeoPointDto"
          },
          "coverage": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logoAsset": {
            "type": "string",
            "nullable": true
          },
          "verification": {
            "type": "string",
            "enum": [
              "NONE",
              "PENDING",
              "VERIFIED",
              "REJECTED"
            ]
          },
          "responseRate": {
            "type": "number",
            "description": "Between 0 and 1."
          },
          "pinned": {
            "type": "boolean"
          },
          "isVerified": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "phone",
          "position",
          "coverage",
          "verification",
          "responseRate",
          "pinned",
          "isVerified"
        ]
      },
      "CreateBrokerDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "INDIVIDUAL",
              "AGENCY"
            ]
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string",
            "nullable": true
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "coverage": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "logoAsset": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "kind",
          "name",
          "phone",
          "latitude",
          "longitude"
        ]
      },
...
      "PropertyDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "brokerId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "LAND",
              "STUDIO",
              "ROOM"
            ]
          },
          "transaction": {
            "type": "string",
            "enum": [
              "RENT",
              "SALE"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "CFA francs, integer."
          },
          "surface": {
            "type": "number",
            "nullable": true
          },
          "rooms": {
            "type": "number",
            "nullable": true
          },
          "position": {
            "$ref": "#/components/schemas/GeoPointDto"
          },
          "neighbourhood": {
            "type": "string"
          },
          "photoAssets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "CLOSED"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "isDiscoverable": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "brokerId",
          "kind",
          "transaction",
          "title",
          "description",
          "price",
          "position",
          "neighbourhood",
          "photoAssets",
          "status",
          "createdAt",
          "isDiscoverable"
        ]
      },
      "UploadPhotoDto": {
        "type": "object",
        "properties": {
          "mimeType": {
            "type": "string",
            "enum": [
              "image/jpeg",
              "image/png",
              "image/webp"
            ]
          },
          "dataBase64": {
            "type": "string",
            "description": "Base64-encoded image bytes. Decoded size must not exceed 163840 bytes — compress before sending."
          }
        },
        "required": [
          "mimeType",
          "dataBase64"
        ]
      },
      "CreatePropertyDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "APARTMENT",
              "HOUSE",
              "LAND",
              "STUDIO",
              "ROOM"
            ]
          },
          "transaction": {
            "type": "string",
            "enum": [
              "RENT",
              "SALE"
            ]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "description": "CFA francs, integer."
          },
          "surface": {
            "type": "number",
            "nullable": true
          },
          "rooms": {
            "type": "number",
            "nullable": true
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "neighbourhood": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "RESERVED",
              "CLOSED"
            ]
          },
          "photoAssets": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "newPhotos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadPhotoDto"
            }
          }
        },
        "required": [
          "kind",
          "transaction",
          "title",
          "price",
          "latitude",
          "longitude",
          "neighbourhood"
        ]
      },
...
      "BrokerListingDto": {
        "type": "object",
        "properties": {
          "broker": {
            "$ref": "#/components/schemas/BrokerDto"
          },
          "distanceMeters": {
            "type": "number"
          },
          "averageRating": {
            "type": "number"
          },
          "reviewCount": {
            "type": "number"
          },
          "availableProperties": {
            "type": "number"
          },
          "score": {
            "type": "number"
          }
        },
        "required": [
          "broker",
          "distanceMeters",
          "averageRating",
          "reviewCount",
          "availableProperties",
          "score"
        ]
      },
      "BrokerSearchResultsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BrokerListingDto"
            }
          },
          "totalCount": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "totalCount",
          "limit",
          "offset"
        ]
      },
      "PropertySearchResultsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyDto"
            }
          },
          "totalCount": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "totalCount",
          "limit",
          "offset"
        ]
      },
      "SearchSuggestionsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}